Remove queue system of packupdater, add persistent installation tracking
Queue system was OK-ish
But it had too many kinks that weren't simple or clean to solve.

By removing it, and adding simple tracking of whether a pack is being installed or not, we can let button statuses persist between window lifecycles, which is going to make this more foolproof and less confusing for the user.

The loading ring was also removed in favor of a installing 3 dot animation text that appears on the button.
The sudden disappearance of the button and long wait times were confusing at best, this at least more clearly tells the user to wait up.